+++ Ink Canvas/App.xaml
@@ -1,37 +1,32 @@
-﻿<Application x:Class="Ink_Canvas.App"
+﻿<Application x:Class="InkCanvasPlus.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:Ink_Canvas"
              StartupUri="MainWindow.xaml"
              xmlns:ui="http://schemas.inkore.net/lib/ui/wpf/modern">
     <Application.Resources>
         <ResourceDictionary>
             <Style TargetType="ui:ScrollViewerEx">
 
-                <EventSetter Event="PreviewMouseWheel" Handler="ScrollViewer_PreviewMouseWheel"/>
-
+                <EventSetter Event="PreviewMouseWheel" Handler="ScrollViewer_PreviewMouseWheel" />
             </Style>
 
             <Style TargetType="ui:FontIcon">
-                <Setter Property="FontSize" Value="20"/>
+                <Setter Property="FontSize" Value="20" />
             </Style>
             <Style TargetType="ui:SymbolIcon">
-                <Setter Property="FontSize" Value="20"/>
+                <Setter Property="FontSize" Value="20" />
             </Style>
 
-
             <ResourceDictionary.MergedDictionaries>
-                
-                <ui:ThemeResources RequestedTheme="Light"/>
-                
-                
-                <ui:XamlControlsResources/>
-                
-                <ResourceDictionary Source="Resources/SeewoImageDictionary.xaml"/>
-                <ResourceDictionary Source="Resources/DrawShapeImageDictionary.xaml"/>
-                <ResourceDictionary Source="Resources/IconImageDictionary.xaml"/>
-                <ResourceDictionary Source="Resources/Styles/Light.xaml"/>
-                
+
+                <ui:ThemeResources RequestedTheme="Light" />
+
+                <ui:XamlControlsResources />
+
+                <ResourceDictionary Source="Resources/SeewoImageDictionary.xaml" />
+                <ResourceDictionary Source="Resources/DrawShapeImageDictionary.xaml" />
+                <ResourceDictionary Source="Resources/IconImageDictionary.xaml" />
+                <ResourceDictionary Source="Resources/Styles/Light.xaml" />
             </ResourceDictionary.MergedDictionaries>
         </ResourceDictionary>
     </Application.Resources>
