--- Ink Canvas/ChangeLogWindow.xaml
@@ -1,51 +1,0 @@
-﻿<Window x:Class="Ink_Canvas.ChangeLogWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Ink_Canvas" WindowStartupLocation="CenterScreen" ResizeMode="NoResize" ShowInTaskbar="False"
-        xmlns:ui="http://schemas.inkore.net/lib/ui/wpf/modern" FontFamily="Microsoft YaHei UI" Topmost="True"
-        mc:Ignorable="d" ui:WindowHelper.UseModernWindowStyle="True" Loaded="Window_Loaded" Closed="Window_Closed"
-        Title="" Height="390" Width="600">
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="50"/>
-            <RowDefinition Height="130"/>
-            <RowDefinition/>
-            <RowDefinition Height="60"/>
-        </Grid.RowDefinitions>
-        <Label Content="Ink Canvas 画板功能更新！" FontFamily="Microsoft YaHei UI" FontSize="26" Margin="20,0"/>
-        <StackPanel Grid.Row="1" Visibility="Collapsed">
-            <TextBlock Name="TextBlockChangeLogTitle" Text="更新日志" FontSize="18" FontWeight="Bold" Margin="35,10,35,0"/>
-            <TextBlock Name="TextBlockChangeLog" Text="修复一堆 Bug。" TextWrapping="Wrap" FontSize="14" Margin="55,5"/>
-            <TextBlock Name="TextBlockChangeLog1" Visibility="Collapsed" Text="修复墨迹识别功能中部分误识别问题。" TextWrapping="Wrap" FontSize="14" Margin="55,5"/>
-            <TextBlock Name="TextBlockSuggestionTitle" Text="建议" FontSize="18" FontWeight="Bold" Margin="35,10,35,0"/>
-            <TextBlock Name="TextBlockSuggestion" Text="建议配置屏幕为：3840×2160 (缩放: 250%)，或1920×1080 (缩放: 125%)，以获得最佳体验。" TextWrapping="Wrap" FontSize="14" Margin="55,5"/>
-        </StackPanel>
-        <Grid Grid.Row="1" Margin="50,0,50,0" Visibility="Visible">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition/>
-                <ColumnDefinition/>
-                <ColumnDefinition/>
-            </Grid.ColumnDefinitions>
-            <Image Source="Resources/InkCanvas.ico"/>
-            <Viewbox Grid.Column="1" Margin="20">
-                <ui:SymbolIcon Symbol="Emoji2" Foreground="#0865B5"/>
-            </Viewbox>
-            <Viewbox Grid.Column="2" Margin="20">
-                <ui:SymbolIcon Symbol="Emoji" Foreground="#169141"/>
-            </Viewbox>
-        </Grid>
-        <ui:SimpleStackPanel VerticalAlignment="Bottom" Grid.Row="2" Margin="50,0">
-            <TextBlock Text="更新日志" FontSize="18" FontWeight="Bold" Margin="0,10,0,0"/>
-            <TextBlock Text="1. 支持多步撤销/重做" TextWrapping="Wrap" FontSize="14" Margin="25,5,0,0"/>
-            <TextBlock Text="2. 添加 PowerPoint 相关功能" TextWrapping="Wrap" FontSize="14" Margin="25,5,0,0"/>
-            <TextBlock Text="3. 提供白板选项" TextWrapping="Wrap" FontSize="14" Margin="25,5,0,0"/>
-            <TextBlock Text="4. 增加双曲线渐近线选项" TextWrapping="Wrap" FontSize="14" Margin="25,5,0,0"/>
-            <!--<TextBlock Text="建议" FontSize="18" FontWeight="Bold" Margin="0,10,0,0"/>
-            <TextBlock Text="建议配置屏幕为：3840×2160 (缩放: 250%)，或1920×1080 (缩放: 125%)，以获得最佳体验（否则可能造成橡皮大小过大或过小等问题，可在设置中调整）。" TextWrapping="Wrap" FontSize="14" Margin="25,5,-10,0"/>-->
-        </ui:SimpleStackPanel>
-        <Label Name="LabelVersion" Grid.Row="3" VerticalContentAlignment="Center" FontSize="12" Margin="20,0"/>
-        <Button Content="确定" Grid.Row="3" HorizontalAlignment="Right" Margin="10" Width="100" FontFamily="Microsoft YaHei UI" Click="Button_Click"/>
-    </Grid>
-</Window>
