+++ Ink Canvas/Controls/ClickSensitiveGrid.xaml
@@ -1,4 +1,4 @@
-﻿<Grid x:Class="Ink_Canvas.ClickSensitiveGrid"
+﻿<Grid x:Class="InkCanvasPlus.Controls.ClickSensitiveGrid"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
