+++ Ink Canvas/CountdownTimerWindow.xaml
@@ -1,11 +1,11 @@
-﻿<Window x:Class="Ink_Canvas.CountdownTimerWindow"
+﻿<Window x:Class="InkCanvasPlus.CountdownTimerWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:Ink_Canvas"
         xmlns:ui="http://schemas.inkore.net/lib/ui/wpf/modern"
-        xmlns:processbars="clr-namespace:Ink_Canvas.ProcessBars"
+        xmlns:inkCanvasPlus="clr-namespace:InkCanvasPlus"
         ui:ThemeManager.RequestedTheme="Light" Topmost="True" Background="Transparent"
         mc:Ignorable="d" WindowStyle="None" AllowsTransparency="True"
         Loaded="Window_Loaded" Closing="Window_Closing" WindowStartupLocation="CenterScreen"
@@ -16,7 +16,7 @@
             <Border MouseMove="WindowDragMove" Visibility="{Binding ElementName=TbCurrentTime, Path=Visibility}" Width="64" Height="15" CornerRadius="8" Background="Gray" Margin="0,0,0,5" HorizontalAlignment="Center" VerticalAlignment="Bottom" />
             <Viewbox x:Name="BigViewController" Margin="20,20,20,20">
                 <Grid Height="180" Width="200">
-                    <processbars:CycleProcessBar x:Name="ProcessBarTime" CurrentValue="0" Width="150" VerticalAlignment="Top"/>
+                    <inkCanvasPlus:CycleProcessBar x:Name="ProcessBarTime" CurrentValue="0" Width="150" VerticalAlignment="Top"/>
                     <ui:SimpleStackPanel MouseMove="WindowDragMove" Orientation="Horizontal" Height="28.5" Margin="0,0,0,25"
                                          VerticalAlignment="Center"
                                          HorizontalAlignment="Center">
