+++ Ink Canvas/CycleProcessBar.xaml
@@ -1,9 +1,8 @@
-﻿<UserControl x:Class="Ink_Canvas.ProcessBars.CycleProcessBar"
+﻿<UserControl x:Class="InkCanvasPlus.CycleProcessBar"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:Ink_Canvas.ProcessBars"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              mc:Ignorable="d" >
     <Viewbox>
         <Grid Width="34" Height="34">
