+++ Ink Canvas/Helpers/Hotkey.cs
@@ -5,7 +5,7 @@ using System.Windows;
 using System.Windows.Input;
 using System.Windows.Interop;
 
-namespace Ink_Canvas
+namespace InkCanvasPlus.Helpers
 {
     static class Hotkey
     {
