+++ Ink Canvas/Helpers/InkRecognizeHelper.cs
@@ -3,7 +3,7 @@ using System.Windows;
 using System.Windows.Ink;
 using System.Windows.Media;
 
-namespace Ink_Canvas.Helpers
+namespace InkCanvasPlus.Helpers
 {
     public class InkRecognizeHelper
     {
