+++ Ink Canvas/MainWindow.xaml
@@ -1,11 +1,11 @@
-<Window Name="window" x:Class="Ink_Canvas.MainWindow"
+<Window Name="window" x:Class="InkCanvasPlus.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Ink_Canvas"
         xmlns:ui="http://schemas.inkore.net/lib/ui/wpf/modern"
-        xmlns:c="clr-namespace:Ink_Canvas.Converter"
+        xmlns:helpers="clr-namespace:InkCanvasPlus.Helpers"
+        xmlns:controls="clr-namespace:InkCanvasPlus.Controls"
         mc:Ignorable="d"
         AllowsTransparency="True"
         WindowStyle="None"
@@ -27,8 +27,8 @@
         Stylus.IsPressAndHoldEnabled="False">
     <!--资源中添加命令-->
     <Window.Resources>
-        <c:VisibilityConverter x:Key="VisibilityConverter" />
-        <c:IsEnabledToOpacityConverter x:Key="IsEnabledToOpacityConverter" />
+        <helpers:VisibilityConverter x:Key="VisibilityConverter" />
+        <helpers:IsEnabledToOpacityConverter x:Key="IsEnabledToOpacityConverter" />
         <RoutedUICommand x:Key="KeyExit" Text=" " />
         <RoutedUICommand x:Key="back_HotKey_Command" Text=" " />
         <RoutedUICommand x:Key="HotKey_Capture" Text=" " />
@@ -1116,15 +1116,15 @@
                         <ui:SimpleStackPanel Name="StackPanelCanvacMain" Margin="10,5,0,5" Spacing="10" Orientation="{Binding ElementName=StackPanelFloatingBar, Path=Orientation}">
                             <Grid Height="24" Width="0">
                             </Grid>
-                            <local:ClickSensitiveGrid Click="SymbolIconCursor_Click" Margin="-10,0,0,0">
+                            <controls:ClickSensitiveGrid Click="SymbolIconCursor_Click" Margin="-10,0,0,0">
                                 <ui:SymbolIcon Symbol="Edit" Foreground="{DynamicResource FloatBarForeground}" />
-                            </local:ClickSensitiveGrid>
+                            </controls:ClickSensitiveGrid>
                             <ui:SymbolIcon Symbol="Setting" Foreground="{DynamicResource FloatBarForeground}" MouseUp="SymbolIconSettings_Click" />
                         </ui:SimpleStackPanel>
                         <ui:SimpleStackPanel Name="StackPanelCanvasControls" Visibility="Visible" Margin="10,5,0,5" Spacing="10" Orientation="{Binding ElementName=StackPanelFloatingBar, Path=Orientation}">
-                            <local:ClickSensitiveGrid Click="SymbolIconCursor_Click">
+                            <controls:ClickSensitiveGrid Click="SymbolIconCursor_Click">
                                 <ui:SymbolIcon Symbol="TouchPointer" Foreground="{DynamicResource FloatBarForeground}" />
-                            </local:ClickSensitiveGrid>
+                            </controls:ClickSensitiveGrid>
                             <ui:SimpleStackPanel Orientation="{Binding ElementName=StackPanelFloatingBar, Path=Orientation}" Spacing="8" Margin="0,1">
                                 <Border Name="BorderPenColorBlack" Background="{Binding ElementName=BtnColorBlack, Path=Background}" MouseUp="BorderPenColorBlack_MouseUp" CornerRadius="100"
                                         Width="{Binding ElementName=BorderPenColorBlack, Path=ActualHeight}">
